Partially working. Memory issues to solve
Reading docs for Gtk it seems like dandling references or something similar may be an issue.

I got to find out the best way to structure the app so that memory does not prematurely released or that references don't become invalid. I think that is an issue for may callback dealing with selections and deletions of those seletions.